pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bgc-viewer"
version = "0.1.0"
description = "A Flask server with static HTML and API endpoints"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "flask>=2.3.0",
    "flask-cors>=4.0.0",
    "python-dotenv>=1.0.1",
    "waitress>=3.0.0",
]
requires-python = ">=3.8.1"
readme = "README.md"
license = {text = "Apache-2.0"}

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
bgc-viewer = "bgc_viewer.app:main"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
